requirements.txt
llamaapi
pandas
st-copy-to-clipboard
streamlit-option-menu